refactor(bridge): 컨트롤러 패키지 이름 변경
