+++ src/main/java/bridge/controller/BridgeGame.java
@@ -1,4 +1,4 @@
-package bridge.service;
+package bridge.controller;
 
 import bridge.BridgeMaker;
 import bridge.BridgeNumberGenerator;
